Add crossover benchmark.
Built to measure rather than assert: L0 swept 4-1024 against dense tensor-core matmul and the naive torch gather; --checkpoint mode for the real layer-8 weights; expected pre-fix to die at high-L0 naive rows.

Rejected: asserting the crossover from a roofline argument; a single-baseline sweep

+++ tests/test_sae_decode.py
@@ -24,7 +24,7 @@ def make_sparse_h(n, n_features, l0, device, dtype=torch.float32, seed=0):
     h = torch.zeros(n, n_features, dtype=dtype)
     for row in range(n):
         cols = torch.randperm(n_features, generator=g)[:l0]
-        h[row, cols] = torch.rand(l0, generator=g) + 0.1
+        h[row, cols] = (torch.rand(l0, generator=g) + 0.1).to(dtype)
     return h.to(device)
